feat: Add @radix-ui/react-dialog component
The commit adds a new component called `@radix-ui/react-dialog` to the project. This component provides a dialog UI that can be used for displaying modal dialogs or pop-up windows. It includes various sub-components such as `Dialog`, `DialogTrigger`, `DialogPortal`, `DialogClose`, `DialogOverlay`, `DialogContent`, `DialogHeader`, `DialogFooter`, `DialogTitle`, and `DialogDescription`. These components can be used to create customizable and accessible dialogs in the application.