feat: Add Input component
This commit adds a new Input component to the project. The Input component is a reusable React component that renders an HTML input element with customizable styles. It accepts various props to configure the input type, class name, and other attributes. This component will be used for user input fields throughout the application.