feat: Add axios.create for backend API configuration
This commit adds the `axios.create` method to configure the backend API in the `axios.ts` file. It sets the `baseURL` to the value of `process.env.API_URL` and includes an `Authorization` header with the value of `process.env.API_KEY`. This configuration will be used for making HTTP requests to the backend API in the application.