feat: Add createTodo and getTodos functions to todo actions
This commit adds the createTodo and getTodos functions to the todo actions in the todo.ts file. The createTodo function sends a POST request to the backend API to create a new todo, while the getTodos function sends a GET request to retrieve all todos. These functions will be used to interact with the backend API for todo-related operations in the application.