feat: Add revalidation after item actions
This commit adds the revalidation of the root path ('/') after each item action in the item.ts file and todo.ts file. The revalidatePath('/') function from the 'next/cache' package is called to trigger the revalidation. This ensures that the data displayed in the application is always up to date after creating, updating, deleting, or moving items or todos.