feat: Add Popover component for displaying popovers
This commit adds the Popover component to the project. The Popover component is a UI component that provides popover functionality using the @radix-ui/react-popover library. It includes a trigger element and content element that can be customized and styled. This component will be used to display popovers in the application.